ddd-x86: refactor
This will hopefully make it easier to work with ipython
